requirements.txt
Flask==3.1.1
flask-cors==6.0.0
gunicorn==21.2.0
python-dotenv==1.1.0

opencv-python-headless==4.11.0.86
pillow==11.2.1
pytesseract==0.3.13
pdfplumber==0.11.6
pdfminer.six==20250327

sentence-transformers==4.1.0
torch==2.7.1
transformers==4.52.4
tokenizers==0.21.1
scikit-learn==1.7.0
numpy>=1.23.0,<2.0
scipy==1.15.3

ultralytics==8.3.155

pinecone==7.0.2

requests==2.32.4
tqdm==4.67.1
python-dateutil==2.9.0.post0
typing_extensions==4.14.0
